# backend/src/challenge/dto/submission-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class SubmissionResponseDto {
  @ApiProperty()
  id: string;

  @ApiProperty()
  status: string;

  @ApiProperty({ type: String, format: 'date-time' })
  createdAt: Date;
}